+++ src/tribler-core/tribler_core/components/metadata_store/category_filter/category.conf
