+++ src/tribler-core/tribler_core/components/metadata_store/category_filter/level2.regex
